# internal/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Docveta API
  version: '1'
  description: 'HTTP API of Docveta, a self-hosted document manager.


    Errors are RFC 9457 problem details (`application/problem+json`). Cookie-authenticated unsafe requests must be same-origin
    (CSRF protection); bearer tokens are exempt.


    This file is the contract: `go test ./internal/api` fails when a route is added or removed without updating it, and the
    integration test validates responses against it.'
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
- url: /
security:
- sessionCookie: []
- bearerToken: []
...
paths:
  /api/v1/admin/audit:
    get:
      summary: Audit log (newest first)
      tags:
      - admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEntry'
        default: &id001
          $ref: '#/components/responses/Problem'
      parameters:
      - name: action
        in: query
        schema: &id002
          type: string
      - name: before
        in: query
        schema:
          type: integer
          format: int64
          description: Entry id to page before
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 500
          default: 100
  /api/v1/admin/settings/oidc:
    get:
      summary: Get oidc settings
      tags:
      - admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OIDCConfig'
        default: *id001
    put:
      summary: Replace oidc settings
      tags:
      - admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OIDCConfig'
        default: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OIDCConfig'
  /api/v1/admin/settings/processing:
    get:
      summary: Get processing settings
      tags:
      - admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingSettings'
        default: *id001
    put:
      summary: Replace processing settings
      tags:
      - admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingSettings'
        default: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessingSettings'
  /api/v1/admin/settings/smtp:
    get:
      summary: Get smtp settings
      tags:
      - admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SMTPConfig'
        default: *id001
    put:
      summary: Replace smtp settings
      tags:
      - admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SMTPConfig'
        default: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SMTPConfig'
  /api/v1/admin/settings/smtp/test:
    post:
      summary: Send a test email to the admin
      tags:
      - admin
      responses:
        '204': &id003
          $ref: '#/components/responses/NoContent'
        default: *id001
  /api/v1/admin/system:
    get:
      summary: System information
      tags:
      - admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemInfo'
        default: *id001
  /api/v1/admin/tasks:
    get:
      summary: Processing tasks and queue stats
      tags:
      - admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                - stats
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TaskView'
                  stats:
                    $ref: '#/components/schemas/QueueStats'
        default: *id001
      parameters:
      - name: status
        in: query
        schema: *id002
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 500
          default: 100
  /api/v1/admin/tasks/{id}/retry:
    post:
      summary: Retry a failed task
      tags:
      - admin
      responses:
        '204': *id003
        default: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /api/v1/admin/users:
    get:
      summary: List users
      tags:
      - admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        default: *id001
    post:
      summary: Create a user
      tags:
      - admin
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
  /api/v1/admin/users/{id}:
    patch:
      summary: Update a user
      tags:
      - admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserUpdate'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      summary: Delete a user (their personal space must be empty)
      tags:
      - admin
      responses:
        '204': *id003
        default: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /api/v1/admin/workers:
    get:
      summary: List OCR workers
      tags:
      - admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Worker'
        default: *id001
    post:
      summary: Register a worker
      tags:
      - admin
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - worker
                - token
                properties:
                  worker:
                    $ref: '#/components/schemas/Worker'
                  token:
                    type: string
                    description: Shown once
        default: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              properties:
                name: *id002
  /api/v1/admin/workers/{id}:
    patch:
      summary: Enable or disable a worker
      tags:
      - admin
      responses:
        '204': *id003
        default: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - enabled
              properties:
                enabled: &id007
                  type: boolean
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      summary: Remove a worker
      tags:
      - admin
      responses:
        '204': *id003
        default: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /api/v1/admin/workers/{id}/rotate-token:
    post:
      summary: Issue a new worker token
      tags:
      - admin
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - token
                properties:
                  token: *id002
        default: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
...
  /api/v1/status:
    get:
      summary: Instance status (public)
      tags:
      - auth
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default: *id001
      security: *id004
...
  /api/v1/admin/settings/office:
    get:
      summary: Document converter (Gotenberg) settings
      tags:
      - admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfficeInfo'
        default:
          $ref: '#/components/responses/Problem'
    put:
      summary: Set the converter address (empty = use DOCVETA_GOTENBERG_URL)
      tags:
      - admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfficeInfo'
        default:
          $ref: '#/components/responses/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfficeSettings'
  /api/v1/admin/settings/office/test:
    post:
      summary: Check that the converter answers
      tags:
      - admin
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        default:
          $ref: '#/components/responses/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfficeSettings'
...
  /api/v1/admin/export:
    get:
      summary: Download everything as a zip (documents, text, notes, files)
      tags:
      - admin
      responses:
        '200':
          description: OK
          content:
            application/zip:
              schema:
                type: string
                format: binary
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-docveta_session
      description: Set by login/setup. Named `docveta_session` when the base URL is plain http.
    bearerToken:
      type: http
      scheme: bearer
      description: Personal access token (`dvt_pat_…`) created in Settings → API tokens.
    workerToken:
      type: http
      scheme: bearer
      description: Worker token (`dvt_wrk_…`) from Administration → Processing.
  responses:
    NoContent:
      description: Done
    Problem:
      description: Error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    AdminUserUpdate:
      properties:
        display_name:
          nullable: true
          type: string
        is_admin:
          nullable: true
          type: boolean
        password:
          nullable: true
          type: string
        status:
          nullable: true
          type: string
      type: object
    AuditEntry:
      properties:
        action:
          type: string
        actor_id:
          format: uuid
          nullable: true
          type: string
        actor_name:
          type: string
        actor_type:
          type: string
        at:
          format: date-time
          type: string
        details:
          additionalProperties:
            nullable: true
          type: object
        id:
          format: int64
          type: integer
        ip:
          type: string
        target_id:
          type: string
        target_type:
          type: string
        user_agent:
          type: string
      required:
      - id
      - at
      - actor_id
      - actor_name
      - actor_type
      - action
      - target_type
      - target_id
      - ip
      - user_agent
      - details
      type: object
...
    Capability:
      properties:
        concurrency:
          type: integer
        engine:
          type: string
        engine_version:
          type: string
        input_mime:
          items:
            type: string
          type: array
        languages:
          items:
            type: string
          type: array
        max_pages_per_task:
          type: integer
        outputs:
          items:
            type: string
          type: array
        tags:
          items:
            type: string
          type: array
        task_type:
          type: string
      required:
      - task_type
      - engine
      type: object
...
    NewUser:
      properties:
        display_name:
          type: string
        email:
          type: string
        is_admin:
          type: boolean
        password:
          nullable: true
          type: string
      type: object
...
    OIDCConfig:
      properties:
        admin_groups:
          items:
            type: string
          type: array
        allowed_groups:
          items:
            type: string
          type: array
        auto_provision:
          type: boolean
        button_label:
          type: string
        client_id:
          type: string
        client_secret:
          nullable: true
          type: string
        disable_password_login:
          type: boolean
        enabled:
          type: boolean
        groups_claim:
          type: string
        has_client_secret:
          type: boolean
        issuer:
          type: string
        link_by_verified_email:
          type: boolean
        scopes:
          items:
            type: string
          type: array
      required:
      - enabled
      - issuer
      - client_id
      - scopes
      - button_label
      - auto_provision
      - allowed_groups
      - admin_groups
      - groups_claim
      - link_by_verified_email
      - disable_password_login
      - has_client_secret
      type: object
...
    Problem:
      properties:
        code:
          type: string
        detail:
          type: string
        errors:
          items:
            properties:
              field:
                type: string
              message:
                type: string
            required:
            - field
            - message
            type: object
          type: array
        extra:
          additionalProperties:
            nullable: true
          type: object
        instance:
          type: string
        status:
          type: integer
        title:
          type: string
        type:
          type: string
      required:
      - type
      - title
      - status
      - code
      type: object
    ProcessingSettings:
      properties:
        archive:
          type: boolean
        fallback_after_minutes:
          type: integer
        lease_seconds:
          type: integer
        max_attempts:
          type: integer
        page_batch_size:
          type: integer
        prefer_tags:
          items:
            type: string
          type: array
        skip_ocr_with_text:
          type: boolean
        worker_offline_minutes:
          type: integer
      required:
      - prefer_tags
      - fallback_after_minutes
      - page_batch_size
      - skip_ocr_with_text
      - archive
      - max_attempts
      - lease_seconds
      - worker_offline_minutes
      type: object
...
    QueueStats:
      properties:
        done_24h:
          type: integer
        failed_24h:
          type: integer
        leased:
          type: integer
        pages_done_24h:
          type: integer
        queued:
          type: integer
      required:
      - queued
      - leased
      - failed_24h
      - done_24h
      - pages_done_24h
      type: object
...
    SMTPConfig:
      properties:
        enabled:
          type: boolean
        from:
          type: string
        has_password:
          type: boolean
        host:
          type: string
        password:
          nullable: true
          type: string
        port:
          type: integer
        security:
          type: string
        username:
          type: string
      required:
      - enabled
      - host
      - port
      - security
      - username
      - from
      - has_password
      type: object
...
    Status:
      properties:
        oidc:
          properties:
            button_label:
              type: string
            enabled:
              type: boolean
          required:
          - enabled
          - button_label
          type: object
        password_login:
          type: boolean
        setup_needed:
          type: boolean
        version:
          type: string
      required:
      - setup_needed
      - version
      - password_login
      - oidc
      type: object
...
    SystemInfo:
      properties:
        database_bytes:
          format: int64
          type: integer
        documents:
          format: int64
          type: integer
        go_version:
          type: string
        pages:
          format: int64
          type: integer
        platform:
          type: string
        queue:
          allOf:
          - $ref: '#/components/schemas/QueueStats'
          nullable: true
        storage_bytes:
          format: int64
          type: integer
        storage_free_bytes:
          format: int64
          type: integer
        users:
          format: int64
          type: integer
        version:
          type: string
        workers_online:
          type: integer
        workers_total:
          type: integer
      required:
      - version
      - go_version
      - platform
      - database_bytes
      - storage_free_bytes
      - documents
      - pages
      - users
      - storage_bytes
      - queue
      - workers_online
      - workers_total
      type: object
    TaskView:
      properties:
        attempt:
          type: integer
        created_at:
          format: date-time
          type: string
        document_id:
          format: uuid
          type: string
        document_title:
          type: string
        finished_at:
          format: date-time
          nullable: true
          type: string
        id:
          format: uuid
          type: string
        last_error:
          type: string
        page_from:
          nullable: true
          type: integer
        page_to:
          nullable: true
          type: integer
        priority:
          type: integer
        status:
          type: string
        type:
          type: string
        worker:
          nullable: true
          type: string
      required:
      - id
      - document_id
      - document_title
      - type
      - status
      - page_from
      - page_to
      - priority
      - attempt
      - worker
      - last_error
      - created_at
      - finished_at
      type: object
...
    User:
      properties:
        created_at:
          format: date-time
          type: string
        date_format:
          type: string
        display_name:
          type: string
        email:
          type: string
        has_password:
          type: boolean
        id:
          format: uuid
          type: string
        is_admin:
          type: boolean
        last_login_at:
          format: date-time
          nullable: true
          type: string
        locale:
          type: string
        status:
          type: string
        theme:
          type: string
        timezone:
          type: string
      required:
      - id
      - email
      - display_name
      - is_admin
      - status
      - locale
      - timezone
      - date_format
      - theme
      - has_password
      - last_login_at
      - created_at
      type: object
    Worker:
      properties:
        active_tasks:
          type: integer
        capabilities:
          items:
            $ref: '#/components/schemas/Capability'
          type: array
        created_at:
          format: date-time
          type: string
        done_24h:
          type: integer
        enabled:
          type: boolean
        failed_24h:
          type: integer
        host:
          type: string
        id:
          format: uuid
          type: string
        last_seen_at:
          format: date-time
          nullable: true
          type: string
        name:
          type: string
        online:
          type: boolean
        protocol_version:
          type: integer
        token_prefix:
          type: string
        version:
          type: string
      required:
      - id
      - name
      - token_prefix
      - enabled
      - protocol_version
      - version
      - host
      - capabilities
      - last_seen_at
      - online
      - active_tasks
      - done_24h
      - failed_24h
      - created_at
      type: object
...
    OfficeInfo:
      type: object
      properties:
        url:
          type: string
        from_env:
          type: boolean
        enabled:
          type: boolean
      required:
      - url
      - from_env
      - enabled
    OfficeSettings:
      type: object
      properties:
        url:
          type: string
...